+++ source/GCPackages/DisableChocolateyCommunitySource/DisableChocolateyCommunitySource.config.ps1
@@ -1,0 +1,11 @@
+Configuration DisableChocolateyCommunitySource {
+    Import-DscResource -ModuleName Chocolatey
+
+    node DisableChocolateyCommunitySource {
+        ChocolateySource disableChocoSource {
+            Ensure = 'Present'
+            Name = 'chocolatey'
+            Disabled = $true
+        }
+    }
+}